+++ bin/build_wasm.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-wasm-pack build --dev